breaking news website/news_hotsport/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.28.0
Django==3.2.10
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
googledrivedownloader==0.4
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
Js2Py==0.71
khtube==0.9.26
newsapi-python==0.2.6
Pillow==8.4.0
pyjsparser==2.7.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
youtube-dl==2021.12.17
zipp==3.6.0
